Added abstract class for DHIS2 responses
